Created the initial license, CoC
Using Apache 2.0 with LLVM exceptions for the license.

Using Contributor Covenant 2.1 for the code of conduct.

The goal of this commit is to create the minimum amount of
structure needed to share a publicly-visible repository that
various stakeholders can practically contribute. That implies
clarifying licensing, etc., for the purposes of getting employer
approval to contribute when that is required. Later
commits would add more interesting project structure and contents.